Sunday changes

+++ app/controllers/posts_controller.rb
@@ -1,34 +1,29 @@
 class PostsController < ApplicationController
+  before_action :authenticate_user!
+
   def index
     sort_attribute = params[:sort]
-    if current_foodie
-      @posts = current_foodie.posts
-    else
-      @posts = []
-    end
     render 'index.html.erb'
   end
 
-  def new
-    render 'new.html.erb'
-  end
-
   def create
-    post = Post.new(
+    @post = Post.new(
+      foodie_id: current_foodie.id,
       text: params[:text],
       address: params[:address]
     )
     if @post.save
       image = Image.new(
-        url: params[:image],
-        post_id: @post.id
-      )
+          url: params[:url],
+          post_id: @post.id
+        )
       image.save
       params[:tag_ids].each do |tag_id|
         post_tag = PostTag.new(post_id: @post.id, tag_id: tag_id)
         post_tag.save
-      flash[:success] = "Post created successfully!"
-      redirect_to "/posts/#{@post.id}"
+        flash[:success] = "Post created successfully!"
+        redirect_to "/posts/#{@post.id}"
+      end
     else
       render 'new.html.erb'
     end
@@ -37,31 +32,41 @@ class PostsController < ApplicationController
   def show
     post_id = params[:id]
     @post = Post.find_by(id: post_id)
-    @comments = @post.comments.all
-    @comment = @post.comments.build
+    @comments = Comment.where(post_id: @post).order("created_at DESC")
     render 'show.html.erb'
   end
 
   def edit
-    post_id = params[:id]
-    @post = Post.find_by(id: post_id)
-    render 'edit.html.erb'
+    if current_foodie
+      post_id = params[:id]
+      @post = Post.find_by(id: post_id)
+      render 'edit.html.erb'
+    else 
+      redirect_to '/login'
+    end
   end
 
   def update
-    post_id = params[:id]
-    post = Post.find_by(id: post_id)
-    post.title = params[:title]
-    post.text = params[:text]
-    flash[:success] = "Post updated!"
-    redirect_to "/posts/#{post.id}"
+    if current_foodie
+      post_id = params[:id]
+      post = Post.find_by(id: post_id)
+      post.text = params[:text]
+      flash[:success] = "Post updated!"
+      redirect_to "/posts/#{post.id}"
+    else 
+      redirect_to '/login'
+    end
   end
 
   def destroy
-    post_id = params[:id]
-    post = Post.find_by(id: post_id)
-    post.destroy
-    flash[:success] = "Post destroyed!"
-    redirect_to '/posts'
+    if current_foodie
+      post_id = params[:id]
+      post = Post.find_by(id: post_id)
+      post.destroy
+      flash[:success] = "Post destroyed!"
+      redirect_to '/posts'
+    else 
+      redirect_to '/login'
+    end
   end
 end
